backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
email-validator
chromadb
sentence-transformers
tf-keras
langchain
langchain-community
langchain-huggingface
huggingface_hub
langchain-openai
openai
python-dotenv
pypdf
rank_bm25
